+++ components/modal/modal.vue
@@ -10,10 +10,7 @@
     ]"
     data-qa="dt-modal"
     :aria-hidden="open"
-    @click.self="close"
-    @keydown.esc="close"
-    @keydown.tab="trapFocus"
-    @after-enter.self="setFocusAfterTransition"
+    v-on="modalListeners"
   >
     <div
       v-if="show && ($slots.banner || bannerTitle)"
@@ -98,6 +95,11 @@
             <icon-close />
           </template>
         </dt-button>
+        <sr-only-close-button
+          v-if="showVisuallyHiddenClose"
+          :visually-hidden-close-label="visuallyHiddenCloseLabel"
+          @close="close"
+        />
       </div>
     </transition>
   </dt-lazy-show>
@@ -110,6 +112,9 @@ import Modal from '@/common/mixins/modal.js';
 import { MODAL_KIND_MODIFIERS, MODAL_SIZE_MODIFIERS } from './modal_constants';
 import { getUniqueString } from '@/common/utils';
 import DtLazyShow from '../lazy_show/lazy_show';
+import { EVENT_KEYNAMES } from '@/common/constants';
+import SrOnlyCloseButtonMixin from '@/common/mixins/sr_only_close_button';
+import SrOnlyCloseButton from '@/common/sr_only_close_button';
 
 /**
  * Modals focus the user’s attention exclusively on one task or piece of information
@@ -123,9 +128,10 @@ export default {
     DtLazyShow,
     DtButton,
     IconClose,
+    SrOnlyCloseButton,
   },
 
-  mixins: [Modal],
+  mixins: [Modal, SrOnlyCloseButtonMixin],
 
   props: {
     /**
@@ -276,10 +282,39 @@ export default {
     return {
       MODAL_KIND_MODIFIERS,
       MODAL_SIZE_MODIFIERS,
+      EVENT_KEYNAMES,
     };
   },
 
   computed: {
+    modalListeners () {
+      return {
+        ...this.$listeners,
+
+        click: event => {
+          (event.target === event.currentTarget) && this.close();
+          this.$emit('click', event);
+        },
+
+        keydown: event => {
+          switch (event.code) {
+            case EVENT_KEYNAMES.esc:
+            case EVENT_KEYNAMES.escape:
+              this.close();
+              break;
+            case EVENT_KEYNAMES.tab:
+              this.trapFocus(event);
+              break;
+          }
+          this.$emit('keydown', event);
+        },
+
+        'after-enter': event => {
+          (event.target === event.currentTarget) && this.setFocusAfterTransition();
+        },
+      };
+    },
+
     open () {
       return `${!this.show}`;
     },
@@ -303,6 +338,14 @@ export default {
         }
       },
     },
+
+    $props: {
+      immediate: true,
+      deep: true,
+      handler () {
+        this.validateProps();
+      },
+    },
   },
 
   methods: {
@@ -317,6 +360,13 @@ export default {
     trapFocus (e) {
       if (this.show) {
         this.focusTrappedTabPress(e);
+      }
+    },
+
+    validateProps () {
+      if (this.hideClose && !this.visuallyHiddenClose) {
+        console.error(`If hideClose prop is true, visuallyHiddenClose and visuallyHiddenCloseLabel props
+        need to be set so the component always includes a close button`);
       }
     },
   },