+++ components/notice/notice.vue
@@ -2,7 +2,6 @@
   <aside
     :class="noticeClass"
     data-qa="notice"
-    v-on="$listeners"
   >
     <dt-notice-icon
       :kind="kind"
@@ -29,6 +28,8 @@
     <dt-notice-action
       :hide-close="hideClose"
       :close-button-props="closeButtonProps"
+      :visually-hidden-close="visuallyHiddenClose"
+      :visually-hidden-close-label="visuallyHiddenCloseLabel"
       v-on="$listeners"
     >
       <!-- @slot Enter a possible action for the user to take, such as a link to another page -->
@@ -42,6 +43,7 @@ import DtNoticeIcon from './notice_icon';
 import DtNoticeContent from './notice_content';
 import DtNoticeAction from './notice_action';
 import { NOTICE_KINDS, NOTICE_ROLES } from './notice_constants.js';
+import SrOnlyCloseButtonMixin from '@/common/mixins/sr_only_close_button';
 
 /**
  * A notice is an informational and assistive message that appears inline with content.
@@ -55,6 +57,8 @@ export default {
     DtNoticeContent,
     DtNoticeAction,
   },
+
+  mixins: [SrOnlyCloseButtonMixin],
 
   props: {
     /**