+++ components/notice/notice_action.vue
@@ -8,23 +8,31 @@
     <dt-button
       v-if="!hideClose"
       ref="closeButton"
+      data-qa="dt-notice-action-close-button"
       size="sm"
       importance="clear"
       circle
       :aria-label="closeButtonProps.ariaLabel ? closeButtonProps.ariaLabel : 'Close'"
       v-bind="closeButtonProps"
-      @click="close"
+      v-on="noticeActionListeners"
     >
       <template #icon>
         <icon-close />
       </template>
     </dt-button>
+    <sr-only-close-button
+      v-if="showVisuallyHiddenClose"
+      :visually-hidden-close-label="visuallyHiddenCloseLabel"
+      @close="close"
+    />
   </div>
 </template>
 
 <script>
 import IconClose from '@dialpad/dialtone/lib/dist/vue/icons/IconClose';
 import DtButton from '../button/button';
+import SrOnlyCloseButtonMixin from '@/common/mixins/sr_only_close_button';
+import SrOnlyCloseButton from '@/common/sr_only_close_button';
 
 export default {
   name: 'DtNoticeAction',
@@ -32,7 +40,10 @@ export default {
   components: {
     IconClose,
     DtButton,
+    SrOnlyCloseButton,
   },
+
+  mixins: [SrOnlyCloseButtonMixin],
 
   props: {
     /**
@@ -62,6 +73,29 @@ export default {
     'close',
   ],
 
+  computed: {
+    noticeActionListeners () {
+      return {
+        ...this.$listeners,
+
+        click: event => {
+          this.close();
+          this.$emit('click', event);
+        },
+      };
+    },
+  },
+
+  watch: {
+    $props: {
+      immediate: true,
+      deep: true,
+      handler () {
+        this.validateProps();
+      },
+    },
+  },
+
   created () {
     if (!this.hideClose && !this.closeButtonProps.ariaLabel) {
       console.error('Invalid props: you must pass in closeButtonProps.ariaLabel if the close button is displayed.');
@@ -81,6 +115,13 @@ export default {
   methods: {
     close () {
       this.$emit('close');
+    },
+
+    validateProps () {
+      if (this.hideClose && !this.visuallyHiddenClose) {
+        console.error('If hideClose prop is true, visuallyHiddenClose and visuallyHiddenCloseLabel props ' +
+          'need to be set so the component always includes a close button');
+      }
     },
   },
 };