+++ components/toggle/toggle.stories.js
@@ -4,6 +4,7 @@ import ToggleDefault from './toggle_default.story.vue';
 import ToggleVariants from './toggle_variants.story.vue';
 import ToggleMdx from './toggle.mdx';
 import DtToggle from './toggle';
+import { TOGGLE_CHECKED_VALUES, TOGGLE_SIZE_MODIFIERS } from '@/components/toggle/toggle_constants';
 
 // Default Prop Values
 export const argsData = {
@@ -27,12 +28,34 @@ export const argTypesData = {
 
   // Props
   checked: {
-    description: 'Used to set the initial state of the toggle',
-    control: 'boolean',
+    description: 'Used to set the initial state of the toggle. Setting "mixed" means it gets the indeterminate state.',
+    control: {
+      type: 'select',
+      options: TOGGLE_CHECKED_VALUES,
+    },
     table: {
       category: 'props',
       type: {
-        summary: 'boolean',
+        summary: 'boolean | "mixed"',
+      },
+      defaultValue: {
+        summary: false,
+      },
+    },
+  },
+
+  size: {
+    description: 'Used to set the size of the toggle',
+    control: {
+      type: 'select',
+      options: Object.keys(TOGGLE_SIZE_MODIFIERS),
+    },
+  },
+
+  id: {
+    table: {
+      defaultValue: {
+        summary: 'generated unique ID',
       },
     },
   },