+++ recipes/comboboxes/combobox_with_popover/combobox_with_popover_default.story.vue
@@ -14,7 +14,8 @@
     :open-with-arrow-keys="openWithArrowKeys"
     :has-suggestion-list="hasSuggestionList"
     :empty-list="emptyList"
-    :modal="modal"
+    :visually-hidden-close="visuallyHiddenClose"
+    :visually-hidden-close-label="visuallyHiddenCloseLabel"
     @escape="onComboboxEscape"
     @highlight="onHighlight"
     @select="onComboboxSelect"
@@ -96,7 +97,6 @@ export default {
     onComboboxSelect (i) {
       this.onSelect(i);
       this.value = this.items[i].number;
-      this.$refs.comboboxWithPopover.closeComboboxList();
     },
 
     onComboboxEscape () {